audio/requirements_madmom.txt
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
fonttools==4.58.4
kiwisolver==1.4.8
madmom @ file:///Users/christophe.thiebaud/github.com/madmom
matplotlib==3.10.3
mido==1.3.3
numpy==2.3.0
packaging==25.0
pillow==11.2.1
pyaml==25.5.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
